requirements.txt
streamlit
yfinance
plotly-express
googlesearch-python